refactor: extract credentialResolver + postInstallDispatcher from install/runner.ts
Closes #975.

Pulls two structural seams out of install/runner.ts:

- credentialResolver.ts owns the NPM credentials pause/resume —
  the in-process pending map, provideCredentials / skipCredentials
  route entries, the waitForCredentials promise the runner awaits,
  and clearPendingCredentials called from abortJob.
- postInstallDispatcher.ts owns the two last-mile reconciliation
  passes the runner fires after every per-template post-deploy has
  run: ensureProxyHosts (#807) and ensureOidcClients (#989).

Both extractions are structural with no behaviour change. The
runner re-exports the prior public surface so the install route
handlers + tests do not have to learn the new file names.

runner.ts drops from 1487 to 1349 LOC. Next step toward the
#975 target shape (job state machine only) is to migrate the
remaining Mustache uses through renderTemplate per #957.
